+++ prototype/__init__.py
@@ -1,3 +1,4 @@
-from .prototype_cpp import arrow_version, process_stream
+from .prototype_cpp import arrow_version
+from .prototype_py import fetch_stream
 
-__all__ = ['arrow_version', 'process_stream']
+__all__ = ["arrow_version", "fetch_stream"]
